Everything
A button was supposedly implemented for dynamic lists adding 5 more
lists with images every click of a button but then some concurrency has
occured with the idea and further development would lead to more
classess to be implemented and would involve putting files into the
external storage. Making it a bit costy for just an Assignment.